Replace UI with the CLI from before
